requirements.txt
appdirs==1.4.3
attrs==17.2.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
idna==2.6
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
websockets==3.4
